third_party/blink/web_tests/fast/shapedetection/shapedetection-empty-input.html

<!DOCTYPE html>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script>

// This test verifies TextDetector.detect() returns an empty list when fed with
// an empty HTMLImageElement.
async_test(function(t) {
  var image = new Image();
  var detector = new TextDetector();
  var tryDetection = function() {
    detector.detect(image)
        .then(detectionResult => {
          assert_equals(detectionResult.length, 0);
          t.done();
        })
        .catch(error => {
          assert_unreached("detect() rejected with error: " + error)
        });
  };
  image.onload = tryDetection;
  image.onerror = tryDetection;
  image.src = "";
}, "Text - detect(empty)");

</script>
